--- .env
@@ -1,1 +1,0 @@
-VITE_APP_BASE_URL=/api
