--- .env.development
@@ -1,1 +1,0 @@
-VITE_APP_BASE_URL=http://testing.biaracmpny.my.id
