+++ src/api/Index.jsx
@@ -1,24 +1,6 @@
-import { baseAPI } from "../config/apiService";
-
-baseAPI.interceptors.request.use(
-  (config) => {
-    const token = localStorage.getItem("token");
-    if (token) {
-      config.headers.Authorization = `Bearer ${token}`;
-    }
-    return config;
-  },
-  (error) => {
-    return Promise.reject(error);
-  }
-);
+import { newBaseAPI, baseAPI } from "../config/apiService";
 
 export const api = {
-  /// API with Token
-  // Login
-  login: (body) => {
-    return baseAPI.post("/v1/users/login", body);
-  },
   getUsers: () => {
     return baseAPI.get("/v1/users");
   },
@@ -28,16 +10,9 @@ export const api = {
   getMahasiswa: () => {
     return baseAPI.get("/v1/mahasiswa");
   },
-  getDosenById: (id) => {
-    return baseAPI.get(`/v1/dosen/${id}`);
-  },
-  signUp: (body) => {
-    return baseAPI.post("/v1/users/signup", body);
-  },
-  createDosen: (body) => {
-    return baseAPI.post("/v1/dosen", body);
-  },
-  createMahasiswa: (body) => {
-    return baseAPI.post("/v1/mahasiswa", body);
+  /// API with Token
+  // Login
+  login: (body) => {
+    return newBaseAPI.post("/v1/users/login", body);
   },
 };
