--- src/utils/index.jsx
@@ -1,3 +1,0 @@
-export const BASE_URL = {
-  API: `${import.meta.env.VITE_APP_BASE_URL}`,
-};
